Moved pegs module from ./src to ./src/pegs
